+++ apps/sim/tools/agiloft/create_record.ts
@@ -1,5 +1,6 @@
 import type { AgiloftCreateRecordParams, AgiloftRecordResponse } from '@/tools/agiloft/types'
-import { buildCreateRecordUrl, executeAgiloftRequest } from '@/tools/agiloft/utils'
+import { buildCreateRecordUrl } from '@/tools/agiloft/utils'
+import { executeAgiloftRequest } from '@/tools/agiloft/utils.server'
 import type { ToolConfig } from '@/tools/types'
 
 export const agiloftCreateRecordTool: ToolConfig<AgiloftCreateRecordParams, AgiloftRecordResponse> =
